requirements.txt
Django==2.2.4

psycopg2-binary==2.8.3
elasticsearch==6.3.0

wagtail==2.7
wagtail-autocomplete==0.3.1
djangorestframework==3.10.3
django-filter==2.2.0

django-extensions==2.1.9

django-colorful==1.3
django-cities-light==3.5.0
